+++ src/app/contact-us/text-container.tsx
@@ -1,4 +1,5 @@
 import styles from '@/styles/app/contact-us/page.module.scss';
+import { ContactList } from '@/app/contact-us/contact-list';
 
 export function TextContainer() {
     return (
@@ -6,6 +7,7 @@ export function TextContainer() {
             <h1>Let's Get in Touch!</h1>
             <p>Have a question or need assistance? Reach out to us via email, phone, or the contact form below. We're eager to assist you.</p>
             <p>Nice hearing from you!</p>
+            <ContactList />
         </div>
     );
 }