+++ src/components/common/top-bar.tsx
@@ -20,7 +20,15 @@ export function TopBar() {
     const profileProvider = useContext(ProfileContext);
 
     useEffect(() => {
-        if (user?.type === AccountType.WORKER && location.pathname.startsWith('/manage')) {
+        if (!user) {
+            return;
+        }
+
+        if (user.type === AccountType.WORKER && location.pathname.startsWith('/manage')) {
+            router.push('/');
+        }
+
+        if (user.type !== AccountType.ADMINISTRATOR && location.pathname === '/contact-us/manage') {
             router.push('/');
         }
 
@@ -93,6 +101,7 @@ export function TopBar() {
                                             <ul>
                                                 <li><a href={'/settings'}>Settings</a></li>
                                                 {user.type !== AccountType.WORKER && <li><Link href={'/manage'}>Manage</Link></li>}
+                                                {user.type === AccountType.ADMINISTRATOR && <li><Link href={'/contact-us/manage'}>Contact management</Link></li>}
                                                 <li><Link href={'/transactions'}>View transactions</Link></li>
                                                 <li onClick={onLogOut}><a href='#'>Logout</a></li>
                                             </ul>